// ui/valset-standardizer-frontend/packages/src/api/generated/valset/schemas/transferFormTemplateViewDTO.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Java 外部估值表标准化工程，基于 Spring Boot 和 Quartz
 * OpenAPI spec version: v1.0.0
 */
import type { YssFormDefinition } from "./yssFormDefinition";
import type { Object } from "./object";

export interface TransferFormTemplateViewDTO {
  /** 模板分类。 */
  category?: string;
  /** 模板说明。 */
  description?: string;
  /** 表单定义。(object) */
  formDefinition?: YssFormDefinition;
  /** 默认初始值。(object) */
  initialValues?: Object;
  /** 模板名称。 */
  name?: string;
  /** 模板版本。 */
  version?: string;
}