// ui/valset-standardizer-frontend/packages/src/api/generated/valset/schemas/transferRuleViewDTO.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Java 外部估值表标准化工程，基于 Spring Boot 和 Quartz
 * OpenAPI spec version: v1.0.0
 */
import type { Object } from "./object";

export interface TransferRuleViewDTO {
  /** 生效开始时间。 */
  effectiveFrom?: string;
  /** 生效结束时间。 */
  effectiveTo?: string;
  /** 是否启用。 */
  enabled?: boolean;
  /** 规则对应表单模板名。 */
  formTemplateName?: string;
  /** 匹配策略。 */
  matchStrategy?: string;
  /** 优先级。 */
  priority?: number;
  /** 规则编码。 */
  ruleCode?: string;
  /** 规则主键。 */
  ruleId?: string;
  /** 规则扩展信息。(object) */
  ruleMeta?: Object;
  /** 规则名称。 */
  ruleName?: string;
  /** 规则版本号。 */
  ruleVersion?: string;
  /** 规则脚本内容。 */
  scriptBody?: string;
  /** 脚本语言。 */
  scriptLanguage?: string;
}